+++ app/src/app/Mapping/Form/TestRunConfigForm.tsx
@@ -797,9 +797,7 @@ export const TestRunConfigForm: React.FunctionComponent<TestRunConfigFormProps> 
               aria-label='Test Run Config SSH Key'
             >
               <FormSelectOption key={0} value='' label='' />
-              {sshKeys?.map((option, index) => (
-                <FormSelectOption key={index} value={option.id} label={option.title} />
-              ))}
+              {sshKeys?.map((option, index) => <FormSelectOption key={index} value={option.id} label={option.title} />)}
             </FormSelect>
             {validatedSSHKeyValue !== 'success' && (
               <FormHelperText>
@@ -864,9 +862,7 @@ export const TestRunConfigForm: React.FunctionComponent<TestRunConfigFormProps> 
               aria-label='Test Run Config Testing Farm Compose'
             >
               <FormSelectOption key={0} value='' label='' />
-              {testingFarmComposes?.map((option, index) => (
-                <FormSelectOption key={index} value={option} label={option} />
-              ))}
+              {testingFarmComposes?.map((option, index) => <FormSelectOption key={index} value={option} label={option} />)}
             </FormSelect>
             {validatedTestingFarmComposeValue !== 'success' && (
               <FormHelperText>